+++ src/main/java/dev/nmarulo/depensaapp/app/shoppinglist/dtos/UpdateShoppingListReq.java
@@ -18,7 +18,7 @@ public class UpdateShoppingListReq {
         
         private Long productId;
         
-        private Integer unitTypeId;
+        private Long unitTypeId;
         
     }
